Memory: fixed bug with wram where banking is not done correctly
There is 2 bugs here, the first is
```
bank = data & 3;
```
where it should be `data & 7`, because there is 8 banks, I think I used
3 because I thought 3 bits.

Second bug (which is the most dangerous) is the wrong order of
operations when getting the correct offset in WRAM for `read_wramx` (or
write).

before:
```
data[0x1000 * bank + addr & 0xFFF]
```
In my head it should do the `*` and `&`, and then finally `+`, but in
reality it is `*`, then `+`, and finally `&`, so we always get writes in
the range of 0x000..=0xFFF. So to fix it we just use brackets (ALWAYS
use brackets and don't depend on order of operations of any language).

after:
```
data[(0x1000 * bank) + (addr & 0xFFF)]
```

NOTE: looking at `CPP reference`, the order is like in Rust, `*` then
`+`, then `&`, but not sure if all languages follow this order.