CPU: edited starting registers to match CGB bootrom
Register `B` value is not zero in case the game is DMG with an official
title, the value of `B` will be the checksum of the title, or `0` if the
title is not recognized.

The value of `B` will also control the `assigned palette`, these
palettes have been extracted.
Source: https://tcrf.net/Notes:Game_Boy_Color_Bootstrap_ROM#Assigned_Palette_Configurations