CPU: implemented STOP instruction, but very sloppy
This is not the actual implementation or how it should work, but its
good enough to pass the tests, and in general it works fine as that it
stops execution and waits for keypresses.

I'm not sure how to handle interrupts here, as some documents say that
IE should be cleared, but even though it is cleared joypad can interrupt
stop mode.

Also, there seem to be some difference between DMG and CGB in this
regard, as in DMG the PPU is stopped, but in CGB the PPU is still
running, but the screen becomes black if it is not mode 3 because the
VRAM is blocked.

Not sure how APU deals with STOP instruction, is sound keeps going on
the tone it was left at? Or stops working.