Front: added command line argument `scale`
It is used to specify the starting size of the window in terms of how
much it is bigger than the original gameboy screen sizes (160 x 144).

I implemented this because I was using https://github.com/daid/GBEmulatorShootout
This framework uses screenshots to measure if the test passes or not,
and so the emulator window size must match the constrains of the normal
gameboy lcd size and not more than that. So when I was using it I was
changing the screensize and compile and change it back again for normal
usage, thought that this would be a good feature for different people
and implemented it.

Maybe another thing to add is the ability to specify starting window
size?