PPU: Fixed sprite `OBJ-to-BG` bit check on overlapping sprites
Pandocs: The priority calculation between sprites disregards OBJ-to-BG
Priority (attribute bit 7). Only the highest-priority nonzero sprite
pixel at any given point is compared against the background. Thus if
a sprite with a higher priority (based on OAM index) but with OBJ-to-BG
Priority turned on overlaps a sprite with a lower priority and a nonzero
background pixel, the background pixel is displayed regardless of the
lower-priority sprite's OBJ-to-BG Priority.

Meaning that if two sprites overlap, only the highest sprite's bit 7
will be checked, the lower will not be checked.

I noticed this in Zelda Orcale of ages, where in the item menu, steps to
reproduce:
- have `Harp of Ages` in the 2nd row
- have any item that opens a menu to choose a type (like the `Seed
shooter`)

If these steps are executed, the `Tune` icon of the harp will be visible
through the menu which should not happen as the menu is above it.

The reason for this is that in the game, the way it hides the `Tune`
icon is by placing a blank sprite that have `3`s as its color (meaning
that they are not 0, transparent) and this sprite is before the `Tune`
sprite and have a background priority turned on. From the above
explanation it should work fine as this blank sprite is the highest
sprite and only its bit 7 should be used to check for background
priority.

What was I doing wrong: The reason it didn't work for my emulator is
because in my emulator, I was using only 1 Fifo and when a sprite is
encountered I'll `MIX` it with the current fifo pixels, these pixels can
be background or sprite pixels (from a previous mix). In the mix I check
the sprite background priority to know if this sprite should be above or
below background. The problem is that if a sprite is below background
(have bit 7 set) then the background pixel will remain and the sprite
pixel will not replace it. Then when a second sprite comes after this
but have a higher index (meaning that it should be below the first
sprite) will compare its bit 7 once again with the background which
should not happen.

This problem happen only in CGB, as in DMG the sprites come in order
(lower X comes first in mixing).

Fix: This is a fix for this problem and also a rewrite for the whole
fifo thing. I used now 2 fifos, one for the sprites and one for the
background. So if two sprites overlap, they will mix together in the
sprite fifo. Then at the end, when pushing a color to the lcd, we will
make the final check between the current winning sprite pixel and the
background pixel, then if sprite should be above background we push its
to the lcd, else we use the background.