Serial: Fixed serial internal timer and passed `boot_sclk_align`
Serial contains an internal timer, in the hardware design the timer is
clocked at 2097152Hz in DMG and normal speed mode, the timer is 8 bits
long. If the serial clock speed bit (bit 1) is 1 (in CGB mode) then the serial will
send a bit on the falling edge of bit 2 of the internal timer. If the
serial clock speed bit is 0 (which is the constant value in DMG) it will
send a bit on the falling edge of bit 7. See the following diagram

```
System clock: ----------------------+
                                    V (increase at 2097152Hz rate)
Serial clock: [7, 6, 5, 4, 3, 2, 1, 0]
               ^              ^
               |--------------|-----------------+ (on the falling edge)
              (0)            (1)   |            |
                                   |            |
                                   |            |
                     (SC.speed) -> |            |
Serial control: [7, 6, 5, 4, 3, 2, 1, 0]        |
               (SC.internal_clock) -> |         |
                                      |         |
                                      \ __AND__ /
                                           |
                                       (send bit)
```

Though in the code I incremented the counter at half the speed and
checked on bits 1 and 6 for super speed (idk what it is called)(0) and normal speed (1).