APU: APU is clocked using the divider now and not from the bus
Tests helped me with this: SameSuite apu/div_write_trigger_10 & div_write_trigger

The APU is clocked from the divider using the 12th bit from the whole
divider or the 4th bit from the upper 8 bits on the falling edge.
Since the APU is clocked by the falling edge, when a user reset the
divider by writing to it, it might clock the APU as well
(div_write_trigger).

There is a special case in the APU where if the divider bit 4 is set and
the APU is turned on by writing to NR52 it skips the first clock for
some reason.