APU: added PCM 12/34 registers and fixed trigger_10 mechanism
Passing: SameSuite apu/div_trigger_volume_10,
div_write_trigger_volume_10, div_write_trigger_volume
These tests deal with DIV events trigger with the APU, but they test the
envelope events instead of the length counter events.

I think the description of the `div_[write]_trigger_10` case is wrong,
the description in the SameSuite test file says:
```
This test verifies that starting the APU while bit 4
of the DIV register is set causes the APU to skip the
first DIV-APU event
```

Here, `skip` might not be the best way to describe it as `skip` means
that it didn't run the event and went to the next one instead but it is
wrong. As when the APU is turned on and the divider bit 4 (5 in double
speed) is high. The APU is `delayed` for one event and not skipped. So
the first event will run but after a delay.

I achieved this by setting the `sequencer_position` to -1 so that it
will take 2 clocks to reach the first event and not 1 clocks as it
should.