Memory: fixed CGB registers memory lock and pass some `misc` tests
The miscs tests in mooneye are for CGB, they are passing now (except for
PPU) but some require bootrom.

Until now I have been hardcoding the values for all hardware if bootrom
is not present to pass the boot tests, but it will get quite complicated
if I continue to do it this way. So I'll try to find another solution.