Cartridge: show errors for sram save/load and no panic on load failure
Errors should be printed instead of silently erroring and clearing the
ram (lost progress in some games).

In `Mizu` we store the RTC data of MBC3 in the same sav file, but other
emulators use `.rtc` file, so when loading another sav file from another
emulator we would not be able to load it correctly because we expect to
find RTC data. So as a solution I made the data optional, if its not
there, then just return zeros.