Printer + Serial: implemented real communication for serial and Printer
This is just an initial implementation of the printer. (printing to the
console)

Gameboy Printer is a hardware device to be connected to the gameboy that
can print images, it was used by the Gameboy Camera, some pokemon games
and Zelda DX.

The printer communicate through packets, the gameboy will send some
packets that contain tile data (similar to PPU tile format) and packets
containing print commands with some parameters (like palette, margin,
etc.).

In order to get the communication working, we fixed some of the
`serial.rs` code, now it works only when GB is a master clock, the next
step is to implement a way for 2 gameboys to communicate.