Printer: used `Rc RefCell for serial device and stored printer in front
Now that we are using `Rc RefCell` we will be able to get a pixel buffer
from the printer and display it in a window or something.

Added Key `P` for handling connecting and disconnecting the printer from
the gameboy. The title of the window will show if the printer is
connected or not.