SaveState: implement `Savable` for arrays and some PPU structs
For now we cannot have the fast implementation for `[u8; N]` (waiting
for `min_specialization` feature).

Also implemented a counter that help compute the size of the savable
without using `Vec`, by implementing a `io::Write` that only compute the
number of bytes written.

Also added `#[inline]` in `Savable` implementations