SaveState + Cartridge: implemented `Savable` to Mappers
The problem is that we cannot have the `Mapper` trait extends `Savable`
as it contains methods that uses generics, which is not supported for
trait objects such as `mapper` in `Cartridge` struct.

This approach was taken from `Plastic`, it only works because the
all mappers structs now are constant in save size, meaning that they do
not have any value on the heap like `Vec`, so it should not be a problem,
its just not really nice implementation.