SaveState: added recovery measure in case of failure in loading
We save the state of the cpu and bus before loading, so that if an error
occurred halfway, we would not get a half loaded state (the emulator
might crash or weird stuff happen) and restore the recovery save state.

Also added a check for the reader after loading, it should be empty and
no more bytes are present.